requirements.txt
asgiref==3.2.7
Django==3.0.6
django-ckeditor==5.9.0
django-hitcount==1.3.2
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.1.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
